migration-add 23-01-2025

+++ database/migrations/2025_01_20_101025_create_wo_contact_details_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wo_contact_details', function (Blueprint $table) {
             $table->id();
             
-            $table->integer('work_order_id')->nullable();
+            $table->unsignedBigInteger('work_order_id')->nullable();
             $table->foreign('work_order_id')->references('id')->on('work_orders')->onDelete('NO ACTION');
             $table->string('wo_client_contact_person')->nullable();;
             $table->string('wo_client_designation')->nullable();;
